Add a new `extra-dispatcher-file' configuration option.
This allows handling other requests to implement additional
course-related functionality by the already running server (instead of
adding more servers for whatever's needed).  The documentation needs
more work though: it refers to `dispatcher/c' which needs to be added,
and there is no real sample code for it.  (My use is too specific to be
used.)