Add `handin-make-page` to tweak the rendered server pages
The default works as it did, so no changes to that --- *except* for the
fact that this uses `xexpr->string` which is very fast compared to the
previous.

Given that the style is done through CSS, parameterizing that could go a
long way, but in some cases (as is in mine), there's additional navbar
content, and additional JS scripting.